Moved number-related files into separate directory
