+++ src/logics/BVLogic.h
@@ -10,7 +10,7 @@ Module: New Logic for BitVector
 
 #include "Logic.h"
 
-#include <common/NumberUtils.h>
+#include <common/numbers/NumberUtils.h>
 
 namespace opensmt {
